requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pdfplumber>=0.10.0
numpy>=1.24.0
torch>=2.0.0,<2.5.0

